Implement binary pinning
We make several changes to make this feasible:

* Move some pin logic into chooseVersion
* Change `pinnedCommit` to be `pin`, a boolean value.  If the bool is true, we implement package-specific pinning behavior (such as forcing the version or commit ID listed)

I'm not entirely sure the chooseVersion call in git.swift is correct (in particular the update flag) but I think this preserves the existing behavior.  I don't believe there is test coverage for this case on the Git side, so it's hard to be sure.