Ignore mistmatches between version and branch.
In common cases, one dependency may be satisfied with a released version (e.g. 1.0, etc.) while another dependency may insist on using a development branch (such as master).  Master is generally going to be later than 1.0, but atpm is more pessimistic.

Change this case to a warning.  Ideally we would verify whether or not master is later than the tag but that is complex, and this solves my problem.